llama-index-integrations/retrievers/llama-index-retrievers-galaxia/pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.codespell]
check-filenames = true
check-hidden = true
# Feel free to un-skip examples, and experimental, you will just need to
# work through many typos (--write-changes and --interactive will help)
skip = "*.csv,*.html,*.json,*.jsonl,*.pdf,*.txt,*.ipynb"

[tool.llamahub]
contains_example = true
import_path = "llama_index.retrievers.galaxia"

[tool.llamahub.class_authors]
GalaxiaRetriever = "rr-yhteys"

[tool.mypy]
disallow_untyped_defs = true
# Remove venv skip when integrated with pre-commit
exclude = ["_static", "build", "examples", "notebooks", "venv"]
ignore_missing_imports = true
python_version = "3.8"

[tool.poetry]
name = "llama-index-retrievers-galaxia"
version = "0.1.0"
description = "llama-index retrievers galaxia integration"
authors = ["Robert Rozanski"]
license = "MIT"
readme = "README.md"
packages = [{include = "llama_index/"}]

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
llama-index-core = "^0.10.0"
python_dateutil = ">= 2.5.3"
setuptools = ">= 21.0.0"
urllib3 = ">= 1.25.3, < 2.1.0"
pydantic = ">= 2"
typing-extensions = ">= 4.7.1"
smabbler-api-client = ">= 2.1.2"

[tool.poetry.group.dev.dependencies]
black = {extras = ["jupyter"], version = "<=23.9.1,>=23.7.0"}
codespell = {extras = ["toml"], version = ">=v2.2.6"}
ipython = "8.10.0"
jupyter = "^1.0.0"
mypy = "0.991"
pre-commit = "3.2.0"
pylint = "2.15.10"
pytest = "7.2.1"
pytest-mock = "3.11.1"
ruff = "0.0.292"
tree-sitter-languages = "^1.8.0"
types-Deprecated = ">=0.1.0"
types-PyYAML = "^6.0.12.12"
types-protobuf = "^4.24.0.4"
types-redis = "4.5.5.0"
types-requests = "2.28.11.8" # TODO: unpin when mypy>0.991
types-setuptools = "67.1.0.0"

[[tool.poetry.source]]
name = "smabbler"
url = "https://pkgs.dev.azure.com/Smabbler/Smabbler.Sdk/_packaging/smabblersdkfeed/pypi/simple/"
priority = "supplemental"
